feature: implementacion de manejo de la contingencia por clasificacion de errores, respondiendo como DTE en formato diferido

+++ internal/domain/dte/contingency/models/contingency_model.go
@@ -39,7 +39,7 @@ type ContingencyReason struct {
 	EndDate           string `json:"fFin"`
 	StartTime         string `json:"hInicio"`
 	EndTime           string `json:"hFin"`
-	ContingencyType   int    `json:"tipoContingencia"`
+	ContingencyType   int8   `json:"tipoContingencia"`
 	ContingencyReason string `json:"motivoContingencia,omitempty"`
 }
